Bluetooth: Mesh: Fix broken LCTL tests
This changes how the `lightness_ctx` works in the tester. The tester
contained a nested `bt_mesh_lightness_srv`, but for the LCTL tests, this
nested struct would not actually be the one in use, because the CTL
Server defines a nested server on its own. This caused problems where
the lightness would be clamped to an undefined lightness range

This commit updates `lightness_ctx` to rather use a pointer to the
associated `bt_mesh_lightness_srv`.

Signed-off-by: Ludvig Jordet <ludvig.jordet@nordicsemi.no>