applications/asset_tracker_v2: Do not use LTO for native_simulator
When building for the native simulator, the build will warn
if we mix lto and non-lto objects, as only the app
is build using lto and not the rest.
This will not fail the build, but cause twister to fail
due to the warning.
So let's disable LTO for native_simulator builds.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>