suit: Increase the default number of mbedTLS keys
SUIT unit tests rely on the automatically assigned volatile key IDs.
Since the key IDs changes with the amount of allowed keys, we should not
change it without regenerating all test payloads.
The new Zephyr brings a Kconfig to control this variable and changes the
default amount from 32 to 16, so the SUIT test platform needs to
override it back to the previous default value.

Ref: NCSDK-31359

Signed-off-by: Tomasz Chyrowicz <tomasz.chyrowicz@nordicsemi.no>